simplify notifictions, use hooks
