+++ src/components/useIncomingWSConcreteNotificationId.ts
@@ -1,11 +1,12 @@
-import { useEffect, useRef, useContext, FC } from "react"
+import { useEffect, useRef, useState } from "react"
 import { WSClient, WebSocketClient } from '../lib/Websocket'
-import { NotificationsContext } from "./NotificationsProvider"
 import { getSessionToken, useUserAuth } from "../hooks/useApollo"
 
-export const WebsocketClient: FC = () => {
-  const {setConcreteNotificationId} = useContext(NotificationsContext)
+type State = number | null
+
+export const useIncomingWSConcreteNotificationId = (): State  => {
   const {sessionState, userId} = useUserAuth()
+  const [concreteNotificationId, setConcreteNotificationId] = useState<State>(null)
   const wsClient = useRef<WebSocketClient | null>(null);
   const close = () => {
     wsClient.current?.close()
@@ -16,23 +17,20 @@ export const WebsocketClient: FC = () => {
     if (sessionState !== 'logged-in' || !userId) {
       close()
       return
-
     }
-    if (!wsClient.current) {
-      const propagateMessage = ({ data }: MessageEvent) => {
+    if (!wsClient.current) {      
+      const url = encodeURI(`${process.env.REACT_APP_WEBSOCKET_URL}?id=${userId}&token=${getSessionToken()}`)
+      wsClient.current = new WSClient(url);
+      wsClient.current.onMessage(({ data }: MessageEvent) => {
         const id = JSON.parse(data).concreteNotificationId
         if (typeof id === "number") {
           setConcreteNotificationId(id)
         }
-      }
-      
-      const url = encodeURI(`${process.env.REACT_APP_WEBSOCKET_URL}?id=${userId}&token=${getSessionToken()}`)
-      wsClient.current = new WSClient(url);
-      wsClient.current.onMessage(propagateMessage);
+      });
       
     }
     return close
   }, [sessionState, userId, setConcreteNotificationId]);
 
-  return null;
+  return concreteNotificationId;
 }
